+++ bitcoind-rpc/src/main/scala/org/bitcoins/rpc/client/common/BitcoindRpcClient.scala
@@ -15,7 +15,7 @@ import org.bitcoins.core.util.{FutureUtil, NetworkUtil}
 import org.bitcoins.core.wallet.fee.FeeUnit
 import org.bitcoins.crypto.{DoubleSha256DigestBE, StringFactory}
 import org.bitcoins.rpc.client.v18.V18AssortedRpc
-import org.bitcoins.rpc.client.v20.{V20AssortedRpc, V20MultisigRpc}
+import org.bitcoins.rpc.client.v20.V20MultisigRpc
 import org.bitcoins.rpc.client.v25.BitcoindV25RpcClient
 import org.bitcoins.rpc.client.v26.BitcoindV26RpcClient
 import org.bitcoins.rpc.config._
@@ -54,8 +54,7 @@ class BitcoindRpcClient(override val instance: BitcoindInstance)(implicit
     with UtilRpc
     with V18AssortedRpc
     with DescriptorRpc
-    with V20MultisigRpc
-    with V20AssortedRpc {
+    with V20MultisigRpc {
 
   private val syncing = new AtomicBoolean(false)
